Update to the last Rust: remove float.

+++ examples/graphics2d/camera.rs
@@ -44,7 +44,7 @@ impl Camera {
                 let ndelta = delta as f32; // between -1.0 and 1.0
 
                 self.curr_zoom *= (1.0 + ndelta * ZOOM_FACTOR);
-                self.scene.zoom(1.0 + (ndelta * ZOOM_FACTOR) as float);
+                self.scene.zoom(1.0 + ndelta * ZOOM_FACTOR);
             }
             event::MouseButtonPressed{x, y, _}  => {
                 self.lastx    = x;
